+++ CHANGELOG.md
@@ -3,6 +3,7 @@
 ### Added
 - Added logo to README.
 - Log2 (plus 1) scaling for Violin Plot
+- Add gene name to expression violin plot.
 
 ### Changed
 - Fix bug from #867 where the view config is temporarily invalid due to null values.